refactor unsupervised.ts
