+++ src/in/web/extension.ts
@@ -18,7 +18,8 @@ import {
     showUpdateReminder,
     startupCreateClient,
 } from '../shared/extension';
-import { addUnsupervisedActions, registerUnsupervisedHandlers } from './unsupervised';
+import { addUnsupervisedActions } from './unsupervised';
+import { registerUnsupervisedHandlers } from '@entry/shared/unsupervised';
 import { registerSendSelectionToNeuro } from '@/editing';
 import { loadIgnoreFiles } from '@/utils/ignore_files';
 import { reregisterAllActions } from '../../rce';
